refactor: modularize streamCTRL_TUI_render.c into sub-components
